refactor: drop auto-retry, offer Reconnect button on network errors
The silent retry loop was unreliable against real WiFi drops — backoff
exhausted before the user could restore the connection, and the resumed
stream often failed silently after context TTL. Simpler: show the normal
error state and add a "Reconnect" button alongside "Retry" that calls
resumeStream() directly, so the user decides when to try.
